viz/requirements.txt
manimce==0.1.0